MAINT: Moving yahoo loader from factory to utils.
