+++ zipline/data/__init__.py
@@ -1,3 +1,4 @@
 from . import loader
+from .loader import load_from_yahoo, load_bars_from_yahoo
 
-__all__ = ['loader']
+__all__ = ['loader', 'load_from_yahoo', 'load_bars_from_yahoo']
